feat: implement admin dashboard and statistics Added AdminDashboardController and AdminDashboardService Created AdminDashboardStats DTO for system metrics Added admin-dashboard templates and styling Updated Bet, Match, and User services with repository methods for statistics Improved admin matches view and navigation

+++ src/main/java/_bg/footballbettingapp/user/repository/UserRepository.java
@@ -19,9 +19,11 @@ public interface UserRepository extends JpaRepository<User, UUID> {
 
     boolean existsByRole(Role role);
 
-    int countByRole(Role role);
+    long countByRole(Role role);
 
     Role role(Role role);
 
-    int countByIsActive(boolean b);
+    long countByIsActive(boolean b);
+
+
 }
